ui: set node fqdn before joining cluster
